faster polling and support ignore level debaunce
ignore level debaunce is used to detect if a pin is changing level.  As an example,  we use it for shock sensor.  In a shock sensor,  a pin will fluctuate while the sensor is active.  Using this flag, we will send a notification on the first pin change detected and when a pin has stabilized after debaunce period.